.claude/worktrees/agent-aad9f10c/backend/requirements.txt
# FastAPI Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Pydantic
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# Authentication & Security
PyJWT[crypto]>=2.11.0
pwdlib[argon2,bcrypt]>=0.2.0
bcrypt>=4.2.0
cryptography>=44.0.0

# Background Jobs
celery>=5.5.0  # 5.5+ required for Python 3.13 support
celery-redbeat>=2.2.0  # Distributed beat scheduler — prevents duplicate task firing on multi-worker deploys
redis>=5.2.0
flower>=2.0.1

# Financial APIs
plaid-python>=17.0.0  # Uses plaid.model.* imports; test before upgrading past v26
yfinance>=0.2.40
finnhub-python>=2.4.20
alpha_vantage>=3.0.0

# Utilities
python-dateutil>=2.9.0
httpx>=0.28.0

# Logging & Monitoring
structlog>=24.4.0
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=2.19.0
prometheus-client>=0.21.0
prometheus-fastapi-instrumentator>=7.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
aiosqlite>=0.20.0

# Development
ruff>=0.8.0
mypy>=1.13.0

# Additional Dev & CI/CD
safety>=3.2.11
pre-commit>=4.0.0

# Email
aiosmtplib>=3.0.0

# Multi-Factor Authentication
pyotp>=2.9.0
qrcode[pil]>=8.0

# Cloud Storage (S3)
boto3>=1.35.0
